+++ packages/babel-generator/test/fixtures/typescript/types-union-intersection/output.js
@@ -1,4 +1,4 @@
 let union: number | null | undefined;
 let intersection: number & string;
-let precedence1: number | (string & boolean);
-let precedence2: (number & string) | boolean;
+let precedence1: number | string & boolean;
+let precedence2: number & string | boolean;