+++ packages/babel-parser/test/error-codes.js
@@ -1,4 +1,5 @@
 import { parse } from "../lib/index.js";
+import { IS_BABEL_8 } from "$repo-utils";
 
 describe("error codes", function () {
   it("raises an error with BABEL_PARSER_SOURCETYPE_MODULE_REQUIRED and reasonCode", function () {
@@ -38,5 +39,24 @@ describe("error codes", function () {
         : "OptionalBindingPattern",
     );
     expect(flowError.message).toBe(tsError.message);
+  });
+  it("Use correct spelling in Babel 8", function () {
+    const code = `
+interface Foo {
+  get foo<T>(): string;
+  set bar<T>(v);
+}
+`;
+    const { errors } = parse(code, {
+      errorRecovery: true,
+      plugins: ["typescript"],
+    });
+    const error = errors[0];
+    expect(error.code).toBe("BABEL_PARSER_SYNTAX_ERROR");
+    expect(error.reasonCode).toBe(
+      IS_BABEL_8()
+        ? "AccessorCannotHaveTypeParameters"
+        : "AccesorCannotHaveTypeParameters",
+    );
   });
 });