+++ packages/babel-plugin-transform-async-to-generator/test/fixtures/async-to-generator/async-iife-with-regenerator/output.js
@@ -1,4 +1,4 @@
-babelHelpers.asyncToGenerator( /*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee() {
+babelHelpers.asyncToGenerator(/*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee() {
   return babelHelpers.regeneratorRuntime().wrap(function _callee$(_context) {
     while (1) switch (_context.prev = _context.next) {
       case 0:
@@ -10,7 +10,7 @@ babelHelpers.asyncToGenerator( /*#__PURE__*/babelHelpers.regeneratorRuntime().ma
     }
   }, _callee);
 }))();
-babelHelpers.asyncToGenerator( /*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee2() {
+babelHelpers.asyncToGenerator(/*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee2() {
   return babelHelpers.regeneratorRuntime().wrap(function _callee2$(_context2) {
     while (1) switch (_context2.prev = _context2.next) {
       case 0:
@@ -23,7 +23,7 @@ babelHelpers.asyncToGenerator( /*#__PURE__*/babelHelpers.regeneratorRuntime().ma
   }, _callee2);
 }))();
 /*#__PURE__*/(function () {
-  var _notIIFE = babelHelpers.asyncToGenerator( /*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee3() {
+  var _notIIFE = babelHelpers.asyncToGenerator(/*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee3() {
     return babelHelpers.regeneratorRuntime().wrap(function _callee3$(_context3) {
       while (1) switch (_context3.prev = _context3.next) {
         case 0:
@@ -40,7 +40,7 @@ babelHelpers.asyncToGenerator( /*#__PURE__*/babelHelpers.regeneratorRuntime().ma
   }
   return notIIFE;
 })();
-/*#__PURE__*/babelHelpers.asyncToGenerator( /*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee4() {
+/*#__PURE__*/babelHelpers.asyncToGenerator(/*#__PURE__*/babelHelpers.regeneratorRuntime().mark(function _callee4() {
   return babelHelpers.regeneratorRuntime().wrap(function _callee4$(_context4) {
     while (1) switch (_context4.prev = _context4.next) {
       case 0: