+++ packages/babel-plugin-transform-async-to-generator/test/fixtures/regression/15978/output.js
@@ -1,7 +1,7 @@
 babelHelpers.asyncToGenerator(function* () {
   let items = [1, 2, 3, 4];
   for (const item of items) {
-    yield f( /*#__PURE__*/function () {
+    yield f(/*#__PURE__*/function () {
       var _ref2 = babelHelpers.asyncToGenerator(function* (x) {
         console.log(item);
       });