+++ packages/babel-plugin-transform-duplicate-named-capturing-groups-regex/test/fixtures/wrapper/basic/output.js
@@ -1,3 +1,3 @@
-"foo".match( /*#__PURE__*/babelHelpers.wrapRegExp(/(.)\1/, {
+"foo".match(/*#__PURE__*/babelHelpers.wrapRegExp(/(.)\1/, {
   double: 1
 }));