+++ packages/babel-plugin-transform-named-capturing-groups-regex/test/fixtures/wrapper/skips-anonymous-capturing-groups/output.js
@@ -1,3 +1,3 @@
-"abba".match( /*#__PURE__*/babelHelpers.wrapRegExp(/(.)(.)\2\1/, {
+"abba".match(/*#__PURE__*/babelHelpers.wrapRegExp(/(.)(.)\2\1/, {
   n: 2
 }));