+++ packages/babel-plugin-transform-react-jsx/test/fixtures/autoImport/after-polyfills-script-not-supported/output.js
@@ -4,6 +4,6 @@ var _reactJsxRuntime = require("react/jsx-runtime");
 require('react-app-polyfill/ie11');
 require('react-app-polyfill/stable');
 const ReactDOM = require('react-dom');
-ReactDOM.render( /*#__PURE__*/_reactJsxRuntime.jsx("p", {
+ReactDOM.render(/*#__PURE__*/_reactJsxRuntime.jsx("p", {
   children: "Hello, World!"
 }), document.getElementById('root'));