+++ packages/babel-plugin-transform-react-jsx/test/fixtures/autoImport/after-polyfills/output.mjs
@@ -4,6 +4,6 @@ import 'react-app-polyfill/ie11';
 import 'react-app-polyfill/stable';
 import ReactDOM from 'react-dom';
 import { jsx as _jsx } from "react/jsx-runtime";
-ReactDOM.render( /*#__PURE__*/_jsx("p", {
+ReactDOM.render(/*#__PURE__*/_jsx("p", {
   children: "Hello, World!"
 }), document.getElementById('root'));